refactor(skills): move skills to ee
