+++ pandasai/agent/state.py
@@ -5,9 +5,10 @@ import uuid
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
-from pandasai.config import Config, ConfigManager, SkillsManager
+from pandasai.config import Config, ConfigManager
 from pandasai.constants import DEFAULT_CHART_DIRECTORY
 from pandasai.data_loader.semantic_layer_schema import is_schema_source_same
+from pandasai.ee.skills.manager import SkillsManager
 from pandasai.exceptions import InvalidConfigError
 from pandasai.helpers.folder import Folder
 from pandasai.helpers.logger import Logger
